1342. Number of Steps to Reduce a Number to Zero
Given a non-negative integer num, return the number of steps to reduce it to zero. If the current number is even, you have to divide it by 2, otherwise, you have to subtract 1 from it.

Runtime: 36 ms, faster than 17.84% of Python3 online submissions for Number of Steps to Reduce a Number to Zero.
Memory Usage: 12.9 MB, less than 100.00% of Python3 online submissions for Number of Steps to Reduce a Number to Zero.