771. Jewels and Stones
You're given strings J representing the types of stones that are jewels, and S representing the stones you have.  Each character in S is a type of stone you have.  You want to know how many of the stones you have are also jewels.

The letters in J are guaranteed distinct, and all characters in J and S are letters. Letters are case sensitive, so "a" is considered a different type of stone from "A".

Runtime: 32 ms, faster than 32.32% of Python3 online submissions for Jewels and Stones.
Memory Usage: 12.9 MB, less than 100.00% of Python3 online submissions for Jewels and Stones.